refactor(TravelogueMultiImageUpload): 내부 로직 hook으로 분리
